refactor: improve code architecture, fix: signin valid schema
